# Speed up pull with larger buffers while writing to layer tempfile